Add SearchFilter component back
