+++ src/components/SearchFilter/SearchFilter.module.scss
@@ -1,0 +1,29 @@
+@import 'styles/uswdsDependencies';
+@import 'styles/sfds/sfdsDependencies';
+
+.searchFilter {
+  background: var(--background-fill);
+  border: 1px solid $light-gray;
+  border-radius: 8px;
+  padding: units(2);
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.15);
+
+  > h3 {
+    margin: 0;
+    font-weight: 500;
+  }
+
+  > .headerContainer {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+}
+
+.checkbox {
+  background-color: transparent;
+}
+
+.buttonContainer {
+  margin-top: 12px;
+}
