+++ src/styles/pages/search.module.scss
@@ -28,30 +28,3 @@
 .noResults {
   display: flex;
 }
-
-.searchFilter {
-  background: var(--background-fill);
-  border: 1px solid $light-gray;
-  border-radius: 8px;
-  padding: units(2);
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.15);
-
-  > h3 {
-    margin: 0;
-    font-weight: 500;
-  }
-
-  > .headerContainer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-}
-
-.checkbox {
-  background-color: transparent;
-}
-
-.buttonContainer {
-  margin-top: 12px;
-}
